+++ resolwe/flow/models/data.py
@@ -5,6 +5,8 @@ import json
 import logging
 
 import jsonschema
+from asgiref.sync import async_to_sync
+from channels.layers import get_channel_layer
 
 from django.contrib.postgres.fields import ArrayField
 from django.contrib.postgres.indexes import GinIndex
@@ -21,7 +23,9 @@ from resolwe.flow.models.utils import (
     validation_schema,
 )
 from resolwe.flow.utils import dict_dot, get_data_checksum, iterate_fields
+from resolwe.observers.consumers import BACKGROUND_TASK_CHANNEL
 from resolwe.observers.decorators import move_to_container
+from resolwe.observers.models import BackgroundTask
 from resolwe.permissions.models import PermissionObject, PermissionQuerySet
 from resolwe.permissions.utils import assign_contributor_permissions, copy_permissions
 
@@ -31,7 +35,6 @@ from .entity import Entity, EntityQuerySet
 from .secret import Secret
 from .storage import Storage
 from .utils import (
-    bulk_duplicate,
     hydrate_input_references,
     hydrate_size,
     render_descriptor,
@@ -206,19 +209,22 @@ class DataQuerySet(BaseQuerySet, PermissionQuerySet):
 
         return obj
 
-    @transaction.atomic
     def duplicate(self, contributor, inherit_entity=False, inherit_collection=False):
-        """Duplicate (make a copy) ``Data`` objects.
+        """Duplicate (make a copy) ``Data`` objects in the background.
 
         :param contributor: Duplication user
         """
-
-        return bulk_duplicate(
-            data=self,
-            contributor=contributor,
-            inherit_entity=inherit_entity,
-            inherit_collection=inherit_collection,
-        )
+        task = BackgroundTask.objects.create(description="Duplicate data")
+        packet = {
+            "type": "duplicate_data",
+            "data_ids": list(self.values_list("pk", flat=True)),
+            "task_id": task.id,
+            "contributor_id": contributor.id,
+            "inherit_entity": inherit_entity,
+            "inherit_collection": inherit_collection,
+        }
+        async_to_sync(get_channel_layer().send)(BACKGROUND_TASK_CHANNEL, packet)
+        return task
 
     @transaction.atomic
     def move_to_collection(self, destination_collection):
@@ -587,13 +593,18 @@ class Data(BaseModel, PermissionObject):
         return bool(self.duplicated)
 
     def duplicate(self, contributor, inherit_entity=False, inherit_collection=False):
-        """Duplicate (make a copy)."""
-        return bulk_duplicate(
-            data=self._meta.model.objects.filter(pk=self.pk),
-            contributor=contributor,
-            inherit_entity=inherit_entity,
-            inherit_collection=inherit_collection,
-        )[0]
+        """Duplicate (make a copy) in the background."""
+        task = BackgroundTask.objects.create(description="Duplicate data")
+        packet = {
+            "type": "duplicate_data",
+            "data_ids": [self.pk],
+            "task_id": task.id,
+            "contributor_id": contributor.id,
+            "inherit_entity": inherit_entity,
+            "inherit_collection": inherit_collection,
+        }
+        async_to_sync(get_channel_layer().send)(BACKGROUND_TASK_CHANNEL, packet)
+        return task
 
     @move_to_container
     def move_to_collection(self, collection):