+++ docs/CHANGELOG.rst
@@ -20,6 +20,8 @@ Changed
 -------
 - **BACKWARD INCOMPATIBLE** Remove the duplicated ``collection`` subfield from the ``entity`` field when
   serializing ``Data`` objects
+- **BACKWARD INCOMPATIBLE:** Replace ``duplicate`` methods on ``Data``, ``Entity`` and
+ ``Collection`` objects with the call to the corresponding background duplicate job.
 - When subscribing to an object for notifications the ``CREATE`` notifications are also
   sent along. This is necessary to receive CREATE notifications triggered from the data
   inside the container on the container.